Biggest update in mewdekos history. Moving from efcore to linqtodb.

+++ src/Mewdeko/Modules/Utility/SlashAi.cs
@@ -109,7 +109,7 @@ public partial class SlashUtility
                 return;
             }
 
-            config.Provider = provider;
+            config.Provider = (int)provider;
             config.Model = model;
             await Service.UpdateConfig(config);
